Moved SequentialGuid
