+++ TransactionalBox.sln
@@ -39,8 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.KeyedInMemoryLock.Tests", "tests\Internals\TransactionalBox.Internals.KeyedInMemoryLock.Tests\TransactionalBox.Internals.KeyedInMemoryLock.Tests.csproj", "{032F6352-5C69-49AF-9275-28E813254591}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.Internals.SequentialGuid", "source\Internals\TransactionalBox.Internals.SequentialGuid\TransactionalBox.Internals.SequentialGuid.csproj", "{57CFF2F1-FB52-4719-A373-A2171A390FEC}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TransactionalBox.End2EndTests", "tests\TransactionalBox.End2EndTests\TransactionalBox.End2EndTests.csproj", "{06A55E8C-9D5B-4E12-B64B-EC59690818C9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Bank", "Bank", "{6BB61EDF-B793-4CD2-94F5-9B11E131845E}"
@@ -121,10 +119,6 @@ Global
 		{032F6352-5C69-49AF-9275-28E813254591}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{032F6352-5C69-49AF-9275-28E813254591}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{032F6352-5C69-49AF-9275-28E813254591}.Release|Any CPU.Build.0 = Release|Any CPU
-		{57CFF2F1-FB52-4719-A373-A2171A390FEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{57CFF2F1-FB52-4719-A373-A2171A390FEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{57CFF2F1-FB52-4719-A373-A2171A390FEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{57CFF2F1-FB52-4719-A373-A2171A390FEC}.Release|Any CPU.Build.0 = Release|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -181,7 +175,6 @@ Global
 		{824310B6-2605-450F-B959-88411A7BD6BE} = {6DEFBC37-33F1-4E42-9A7E-2752E9AD3187}
 		{FFA83AFD-4303-4447-99DD-6FEEEB9FC2C4} = {824310B6-2605-450F-B959-88411A7BD6BE}
 		{032F6352-5C69-49AF-9275-28E813254591} = {824310B6-2605-450F-B959-88411A7BD6BE}
-		{57CFF2F1-FB52-4719-A373-A2171A390FEC} = {9C70A8B9-322F-49D5-AF08-889B6BBB161C}
 		{06A55E8C-9D5B-4E12-B64B-EC59690818C9} = {6DEFBC37-33F1-4E42-9A7E-2752E9AD3187}
 		{6BB61EDF-B793-4CD2-94F5-9B11E131845E} = {D018E603-5AE9-4B1B-BF33-04EB0B5D6913}
 		{580972A5-B9DC-49B0-8CFB-28412E3FC5C5} = {6BB61EDF-B793-4CD2-94F5-9B11E131845E}
