--- source/Internals/TransactionalBox.Internals.SequentialGuid/TransactionalBox.Internals.SequentialGuid.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-	<ItemGroup>
-		<InternalsVisibleTo Include="TransactionalBox" />
-	</ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" />
-  </ItemGroup>
-
-</Project>
